+++ apps/mobile/app/dashboard/(tabs)/_layout.tsx
@@ -2,7 +2,13 @@ import React, { useLayoutEffect } from "react";
 import { Tabs, useNavigation } from "expo-router";
 import { StyledTabs } from "@/components/navigation/tabs";
 import { useColorScheme } from "@/lib/useColorScheme";
-import { ClipboardList, Home, Settings, Tag } from "lucide-react-native";
+import {
+  ClipboardList,
+  Highlighter,
+  Home,
+  Settings,
+  Tag,
+} from "lucide-react-native";
 
 export default function TabLayout() {
   const { colors } = useColorScheme();
@@ -42,6 +48,13 @@ export default function TabLayout() {
         options={{
           title: "Tags",
           tabBarIcon: ({ color }) => <Tag color={color} />,
+        }}
+      />
+      <Tabs.Screen
+        name="highlights"
+        options={{
+          title: "Highlights",
+          tabBarIcon: ({ color }) => <Highlighter color={color} />,
         }}
       />
       <Tabs.Screen
