+++ apps/mobile/app/dashboard/_layout.tsx
@@ -58,14 +58,6 @@ export default function Dashboard() {
         }}
       />
       <Stack.Screen
-        name="highlights"
-        options={{
-          headerTitle: "",
-          headerBackTitle: "Back",
-          headerTransparent: true,
-        }}
-      />
-      <Stack.Screen
         name="bookmarks/[slug]/index"
         options={{
           headerTitle: "",
